+++ src/main/java/com/radieske/reservasapi/repository/ProvedorRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.radieske.reservasapi.model.Provedor;
 
-public interface ProvedorRepository extends JpaRepository<Provedor, Long>
+public interface ProvedorRepository extends JpaRepository<Provedor, Integer>
 {
 	
 }
