+++ src/main/java/com/radieske/reservasapi/repository/UsuarioRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.radieske.reservasapi.model.Usuario;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Long>
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer>
 {
 	Optional<Usuario> findByUsuario(String usuario);
 }
