+++ src/main/java/com/radieske/reservasapi/service/impl/UsuarioServiceImpl.java
@@ -29,7 +29,7 @@ public class UsuarioServiceImpl implements UsuarioService
 	}
 
 	@Override
-	public Optional<Usuario> findById(Long id)
+	public Optional<Usuario> findById(Integer id)
 	{
 		return usuarioRepository.findById(id);
 	}
@@ -41,7 +41,7 @@ public class UsuarioServiceImpl implements UsuarioService
 	}
 
 	@Override
-	public void deleteById(Long id)
+	public void deleteById(Integer id)
 	{
 		usuarioRepository.deleteById(id);
 	}
